+++ Sources/MapboxDirections/MapMatching/MatchResponse.swift
@@ -4,7 +4,7 @@ class MatchResponse: Codable {
     var code: String
     var message: String?
     var matches : [Match]?
-    var tracepoints: [Tracepoint?]?
+    var tracepoints: [Match.Tracepoint?]?
     
     private enum CodingKeys: String, CodingKey {
         case code
@@ -17,12 +17,14 @@ class MatchResponse: Codable {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         code = try container.decode(String.self, forKey: .code)
         message = try container.decodeIfPresent(String.self, forKey: .message)
-        tracepoints = try container.decodeIfPresent([Tracepoint?].self, forKey: .tracepoints)
+        tracepoints = try container.decodeIfPresent([Match.Tracepoint?].self, forKey: .tracepoints)
         matches = try container.decodeIfPresent([Match].self, forKey: .matches)
         
-        if let points = self.tracepoints {
-            matches?.forEach {
-                $0.tracepoints = points
+        if let tracepoints = self.tracepoints, let matches = matches {
+            for match in matches {
+                // TODO: Filter on matchings_index.
+                // TODO: Push tracepoints down to individual legs to reflect waypoint_index.
+                match.tracepoints = tracepoints
             }
         }
     }