+++ Sources/MapboxDirections/MapMatching/Tracepoint.swift
@@ -1,40 +1,44 @@
 import Foundation
 import CoreLocation
 
-/**
- A `Tracepoint` represents a location matched to the road network.
- */
-public class Tracepoint: Waypoint {
+public extension Match {
     /**
-     Number of probable alternative matchings for this tracepoint. A value of zero indicates that this point was matched unambiguously.
+     A tracepoint represents a location matched to the road network.
      */
-    public let countOfAlternatives: Int
-    
-    private enum CodingKeys: String, CodingKey {
-        case countOfAlternatives = "alternatives_count"
-    }
-    
-    init(coordinate: CLLocationCoordinate2D, countOfAlternatives: Int, name: String?) {
-        self.countOfAlternatives = countOfAlternatives
-        super.init(coordinate: coordinate, name: name)
-    }
-    
-    required public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        countOfAlternatives = try container.decode(Int.self, forKey: .countOfAlternatives)
-        try super.init(from: decoder)
-    }
-    
-    public override func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(countOfAlternatives, forKey: .countOfAlternatives)
-        try super.encode(to: encoder)
+    struct Tracepoint: Matchpoint, Equatable {
+        // MARK: Positioning the Waypoint
+        
+        /**
+         The geographic coordinate of the waypoint, snapped to the road network.
+         */
+        public var coordinate: CLLocationCoordinate2D
+        
+        /**
+         The straight-line distance from this waypoint to the corresponding waypoint in the `RouteOptions` or `MatchOptions` object.
+         
+         The requested waypoint is snapped to the road network. This property contains the straight-line distance from the original requested waypoint’s `DirectionsOptions.Waypoint.coordinate` property to the `coordinate` property.
+         */
+        public var correction: CLLocationDistance
+        
+        // MARK: Determining the Degree of Confidence
+        
+        /**
+         Number of probable alternative matchings for this tracepoint. A value of zero indicates that this point was matched unambiguously.
+         */
+        public var countOfAlternatives: Int
     }
 }
 
-extension Tracepoint { //Equatable
-    public static func ==(lhs: Tracepoint, rhs: Tracepoint) -> Bool {
-        let superEquals = (lhs as Waypoint == rhs as Waypoint)
-        return superEquals && lhs.countOfAlternatives == rhs.countOfAlternatives
+extension Match.Tracepoint: Codable {
+    private enum CodingKeys: String, CodingKey {
+        case coordinate = "location"
+        case correction = "distance"
+        case countOfAlternatives = "alternatives_count"
+    }
+}
+
+extension Match.Tracepoint: CustomStringConvertible {
+    public var description: String {
+        return "<latitude: \(coordinate.latitude); longitude: \(coordinate.longitude)>"
     }
 }