+++ Sources/MapboxDirections/RouteLeg.swift
@@ -37,8 +37,9 @@ open class RouteLeg: Codable {
      */
     public required init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        source = try container.decodeIfPresent(Waypoint.self, forKey: .source)
-        destination = try container.decodeIfPresent(Waypoint.self, forKey: .destination)
+        
+        source = try container.decodeIfPresent(Route.Waypoint.self, forKey: .source)
+        destination = try container.decodeIfPresent(Route.Waypoint.self, forKey: .destination)
         steps = try container.decode([RouteStep].self, forKey: .steps)
         name = try container.decode(String.self, forKey: .name)
         distance = try container.decode(CLLocationDistance.self, forKey: .distance)
@@ -85,7 +86,7 @@ open class RouteLeg: Codable {
      
      This property is set to `nil` if the leg was decoded from a JSON RouteLeg object.
      */
-    public var source: Waypoint?
+    public var source: Route.Waypoint?
 
     /**
      The endpoint of the route leg.
@@ -94,7 +95,7 @@ open class RouteLeg: Codable {
      
      This property is set to `nil` if the leg was decoded from a JSON RouteLeg object.
      */
-    public var destination: Waypoint?
+    public var destination: Route.Waypoint?
     
     // MARK: Getting the Steps Along the Leg